backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.7
spacy==3.7.2
yake==0.4.8
scikit-learn==1.3.2
nltk==3.8.1
python-dotenv==1.0.0
llama-index==0.9.15
anthropic==0.5.0
reportlab==4.0.8